Access database from $XDG_DATA_HOME/nvim/scmru.sqlite3
fallback to ~/.cache directory if the former does not exist

+++ lua/mru/modify.lua
@@ -3,8 +3,12 @@ local scm = require("mru.scm")
 
 local M = {}
 
-function M.init()
-    local conn = sqlite.open(vim.g.mru_db_path .. "/mru.db")
+function M.init_db()
+    if vim.g.mru_db_path == nil then
+        return
+    end
+
+    local conn = sqlite.open(vim.g.mru_db_path)
     conn:exec("CREATE TABLE IF NOT EXISTS mru_list(path TEXT PRIMARY KEY, root TEXT, ts INTEGER, freq INTEGER);")
     conn:close()
 end
@@ -15,6 +19,10 @@ local function file_exists(name)
 end
 
 function M.add(path)
+    if vim.g.mru_db_path == nil
+        then return
+    end
+
     if path == nil or path == "" then
         path = vim.api.nvim_buf_get_name(0)
     end
@@ -22,7 +30,7 @@ function M.add(path)
         return
     end
 
-    local conn = sqlite.open(vim.g.mru_db_path .. "/mru.db")
+    local conn = sqlite.open(vim.g.mru_db_path)
 
     local ts = 0
     do
@@ -58,10 +66,14 @@ function M.add(path)
 end
 
 function M.del(path)
+    if vim.g.mru_db_path == nil then
+        return
+    end
+
     if path == nil then
         path = vim.api.nvim_buf_get_name(0)
     end
-    local conn = sqlite.open(vim.g.mru_db_path .. "/mru.db")
+    local conn = sqlite.open(vim.g.mru_db_path)
     conn:exec("DELETE from mru_list WHERE path LIKE '%" .. path .. "';")
     conn:close()
 end
